Describe limit on search-based user queries (#2335)
* Describe limit on search-based user queries

When paginating through the users matching a search-based query, at most 50000 results can be viewed.  The current behavior is a bit rough around the edges (see OKTA-189268) so I describe it as it currently works, where an error is returned if the user tries to follow the "next" link after 50000 results.  We may need to touch up this description after that issue is addressed.

I deliberately avoid mentioning the behavior when fewer than 50000 results have been seen but the "next" link tries to go past 50000.  That cannot arise with the default "limit" of 200 per page (or any other size that evenly divides 50000), and the fix for OKTA-189268 should make it so that pagination always hits 50000 results exactly before yielding an error when requesting the "next" page.

* Change directional doublequotes to (") characters

* Reword per feedback